requirements.txt
fastapi[standard]
psycopg2
wheel
google-analytics-data
pandas
